Working refactor
